+++ src/features/auth/Login.jsx
@@ -4,7 +4,7 @@ import { useForm } from 'react-hook-form'
 import { yupResolver } from '@hookform/resolvers/yup'
 import { loginSchema } from './formSchema'
 
-import { useLogin } from '../../hooks/useLogin'
+import { useLogin } from '../../hooks/auth/useLogin'
 
 function Login() {
   const {
