+++ src/features/auth/Signup.jsx
@@ -3,7 +3,7 @@ import { Link, useNavigate } from 'react-router-dom'
 import { useForm } from 'react-hook-form'
 import { yupResolver } from '@hookform/resolvers/yup'
 import { signUpSchema } from './formSchema'
-import { useSignup } from '../../hooks/useSignup'
+import { useSignup } from '../../hooks/auth/useSignup'
 
 function Signup() {
   const {
