--- src/hooks/useCollection.js
@@ -1,27 +1,0 @@
-import { useState, useEffect } from 'react'
-import { db } from '../config/firebase.ts'
-
-import { collection, onSnapshot } from 'firebase/firestore'
-
-export const useCollection = (c) => {
-  const [documents, setDocuments] = useState(null)
-
-  useEffect(() => {
-    let ref = collection(db, c)
-
-    const unsub = onSnapshot(ref, (snapshot) => {
-      let results = []
-      snapshot.docs.forEach((doc) => {
-        results.push({
-          ...doc.data(),
-          id: doc.id,
-        })
-      })
-      setDocuments(results)
-
-      return () => unsub()
-    })
-  }, [])
-
-  return { documents }
-}
